requirements.txt
flask
ultralytics
opencv-python-headless
pytesseract
Pillow
numpy
flask-cors